+++ R/Sims.R
@@ -1,13 +1,15 @@
-#' @title A dynamics graphic to illustrate the concept of confidence intervals for a mean.
+#' @title A dynamic graph to illustrate the concept of confidence intervals for a mean.
 #' 
-#' @description This function demonstrates the concept of confidence regions by drawing a large number of random samples from a known normal distribution, computing a confidence interval for each sample, and ploting those confidence intervals.  Slider bars then let the user change the level of confidence, the sample size, or the type of confidence region (interval or bound) to see how that effects the confidence interval widths (margin-of-error) and capture probability.
+#' @description This function demonstrates the concept of confidence regions by drawing a large number of random samples from a known normal distribution, computing a confidence interval for each sample, and plotting those confidence intervals.  Sliders then let the user change the level of confidence, the sample size, or the type of confidence region (interval or bound) to see how that effects the confidence interval widths (margin-of-error) and capture probability.
 #' 
-#' @param reps A single numeric indicating the number of replicate samples to draw from the population
-#' @param method A single string indicating whether to make confidence intervals using a normal (\code{="z"}) or t (\code{="t"}) distribution
-#' @param mu A single numeric indicating the mean of the known normal distribution
-#' @param sigma A single numeric indicating the standard deviation of the known normal distribution
+#' @details If the user is using RStudio and the \pkg{manipulate} package is installed then the dynamic graph is produced in the \dQuote{Plots} pane of RStudio.  The plot controls may be accessed through the \dQuote{gear} that is in the upper-left corner of the plot.  If the user is not using RStudio or the \pkg{manipulate} package is not installed, then an attempt is made to produce the dynamic graph with Tcl/Tk using the functions in the \pkg{relax} package.
 #' 
-#' @return None, but a dynamic graphic with slider bars is produced.
+#' @param reps A single numeric that indicates the number of replicate samples to draw from the population
+#' @param method A single string that indicates whether to make confidence intervals using a normal (\code{="Z"}) or t (\code{="t"}) distribution
+#' @param mu A single numeric that indicates the mean of the known normal distribution
+#' @param sigma A single numeric that indicates the standard deviation of the known normal distribution
+#' 
+#' @return None, but a dynamic graphic with sliders is produced.
 #' 
 #' @keywords misc dynamic
 #' 
@@ -144,18 +146,20 @@ iCISimPlot <- function(n,conf,alternative=c("two.sided","less","greater"),
 
 
 
-#' A dynamic graphic to illustrate the Central Limit Theorem.
+#' @title A dynamic graph to illustrate the Central Limit Theorem.
 #' 
-#' A dynamic graphic to illustrate the Central Limit Theorem.  The user can change the population distribution that is sampled from and the sample size.
+#' @description A dynamic graph to illustrate the central limit theorem (CLT).  The user can change the population distribution that is sampled from and the sample size.
 #' 
-#' This function produces two graphics.  The left-most graphic is a histogram of the individuals in the population and the right-most graphic is a histogram of the simulated sampling distribution (i.e., means from the multiple samples).  The right-most graphic may include a normal distribution density curve if the \code{incl.norm} argument is used.
+#' @details This function produces two graphics.  The left graphic is a histogram of the individuals in the population and the right graphic is a histogram of the simulated sampling distribution (i.e., means from the multiple samples).  The right graphic may include a normal distribution density curve with \code{incl.norm=TRUE}.
 #' 
-#' The two graphics are dynamically controlled by three slider bars.  The first two slider bars control the shape parameters of the beta distribution used to model the population distribution that will be sampled from.  The beta distribution allows for a wide variety of shapes for the population distribution.  The last slider bar controls the size of each sample taken from the population distribution.  The slider bars can be used to detect how changes in the shape of the population and the size of the sample effect the shape, center, and dispersion of the sampling distribution.
+#' The two graphs are dynamically controlled by three sliders.  The first two sliders control the shape parameters of the beta distribution used to model the population distribution sampled from.  A wide variety of shapes may be created with these two shape parameters.  The third slider controls the size of each sample taken from the population distribution.  The sliders may be used to detect how changes in the shape of the population and the size of the sample effect the shape, center, and dispersion of the sampling distribution.
 #' 
-#' @param reps Number of samples to take from the population distribution and means to calculate
-#' @param incl.norm A logical indicating whether a normal density curve should be superimposed on the sampling distribution
+#' @details If the user is using RStudio and the \pkg{manipulate} package is installed then the dynamic graph is produced in the \dQuote{Plots} pane of RStudio.  The plot controls may be accessed through the \dQuote{gear} that is in the upper-left corner of the plot.  If the user is not using RStudio or the \pkg{manipulate} package is not installed, then an attempt is made to produce the dynamic graph with Tcl/Tk using the functions in the \pkg{relax} package.
 #' 
-#' @return None, but a dynamic graphic with slider bars will be produced.
+#' @param reps Number of resamples to take from the population distribution and means to calculate
+#' @param incl.norm A logical that indicates whether a normal density curve should be superimposed on the sampling distribution
+#' 
+#' @return None, but a dynamic graph with sliders is produced.
 #' 
 #' @keywords misc dynamic
 #' 
@@ -166,25 +170,27 @@ iCISimPlot <- function(n,conf,alternative=c("two.sided","less","greater"),
 #' 
 #' @export
 cltSim <- function(reps=1000,incl.norm=FALSE) {
+  ## Trying to fix "no visible bindings" problem for Check
+  n <- shape1 <- shape2 <- NULL
   if (iCheckRStudio() & requireNamespace("manipulate",quietly=TRUE)) {
     rerand <- TRUE
     manipulate::manipulate(
       {
         if (rerand) set.seed(sample(1:10000))
-        iCLTSimPlot(n,sh1,sh2,reps,incl.norm)
+        iCLTSimPlot(n,shape1,shape2,reps,incl.norm)
       },
       n=manipulate::slider(10,100,step=5,initial=10),
-      sh1=manipulate::slider(1,20,step=1,label="Shape 1 (alpha)"),
-      sh2=manipulate::slider(1,20,step=1,label="Shape 2 (beta)"),
+      shape1=manipulate::slider(1,20,step=1,label="Shape 1 (alpha)"),
+      shape2=manipulate::slider(1,20,step=1,label="Shape 2 (beta)"),
       rerand=manipulate::button("Rerandomize")
     ) # end manipulate
   } else {
     ## internal referesher function
     iCLTrefresh <- function(...) {
       n <- relax::slider(no=1)
-      sh1 <- relax::slider(no=2)
-      sh2 <- relax::slider(no=3)
-      iCLTSimPlot(n,sh1,sh2,reps,incl.norm)
+      shape1 <- relax::slider(no=2)
+      shape2 <- relax::slider(no=3)
+      iCLTSimPlot(n,shape1,shape2,reps,incl.norm)
     } # end iCLTrefresh internal function
     if (iChk4Namespace("relax")) {
       relax::gslider(iCLTrefresh,prompt=TRUE,hscale=2,
@@ -204,19 +210,19 @@ cltSim <- function(reps=1000,incl.norm=FALSE) {
   }
 }
 
-iCLTSimPlot <- function(n,sh1,sh2,reps,incl.norm) {
+iCLTSimPlot <- function(n,shape1,shape2,reps,incl.norm) {
   ## Create a population with a beta distribution
   x <- seq(0,0.95,0.05)
-  y <- round(1000*stats::dbeta(x,sh1,sh2),0)
+  y <- round(1000*stats::dbeta(x,shape1,shape2),0)
   x <- rep(x,y)
   # make the label
   title.popn <- substitute(paste(mu,"=",muval,", ",sigma,"=",sigmaval),
-                list(muval=formatC(sh1/(sh1+sh2),format="f",digits=2),
-                     sigmaval=formatC(sqrt((sh1*sh2)/(((sh1+sh2)^2)*(sh1+sh2+1))),
+                list(muval=formatC(shape1/(shape1+shape2),format="f",digits=2),
+                     sigmaval=formatC(sqrt((shape1*shape2)/(((shape1+shape2)^2)*(shape1+shape2+1))),
                                       format="f",digits=3)))
   
   ## Construct random samples from pop and their mean
-  rnd.reps <- matrix(stats::rbeta(n*reps,sh1,sh2),nrow=n)
+  rnd.reps <- matrix(stats::rbeta(n*reps,shape1,shape2),nrow=n)
   rnd.mns <- apply(rnd.reps,2,mean)
   # compute mean and SE of sampling dist
   mn.mns <- mean(rnd.mns)
@@ -235,8 +241,8 @@ iCLTSimPlot <- function(n,sh1,sh2,reps,incl.norm) {
   graphics::mtext(title.popn,3,line=0.1)
   graphics::mtext("Frequency of Individuals",2,line=0.1)
   # Plot the sampling distribution
-  # Set limits differently depending on sh1 and sh2
-  if ((sh1+sh2)==2) tmp <- c(0.05,0.95)
+  # Set limits differently depending on shape1 and shape2
+  if ((shape1+shape2)==2) tmp <- c(0.05,0.95)
     else tmp <- c(0.10,0.95)
   mns.hist <- graphics::hist(rnd.mns,freq=FALSE,breaks=15,
                              xlim=stats::quantile(x,tmp),
@@ -260,18 +266,20 @@ iCLTSimPlot <- function(n,sh1,sh2,reps,incl.norm) {
 
 
 
-#' Dynamic simulation of power calculation in 1-sample mean problems.
+#' @title Dynamic simulation of power calculation in 1-sample mean problems.
 #' 
-#' This function plots the null and actual distributions, highlights the rejection region and the power region, and allows the user to manipulate the effect size, standard deviation, sample size, and alpha to determine the effect on power.
+#' @description This function plots the null and actual distributions, highlights the rejection region and the power region, and allows the user to dynamically manipulate the effect size, standard deviation, sample size, and alpha to determine the effect of these parameters on power.
 #' 
-#' @param mu0 A single numeric that is the null hypothesized mean.
-#' @param s.mua A single numeric that is the starting value for the actual population mean.
-#' @param s.sigma A single numeric that is the starting value for the actual population standard deviation.
-#' @param s.n A single numeric that is the starting value for the sample size.
-#' @param s.alpha A single numeric that is the starting value for alpha.
-#' @param lower.tail A single logical indicating if the rejection region is into the lower tail or not.
+#' @details If the user is using RStudio and the \pkg{manipulate} package is installed then the dynamic graph is produced in the \dQuote{Plots} pane of RStudio.  The plot controls may be accessed through the \dQuote{gear} that is in the upper-left corner of the plot.  If the user is not using RStudio or the \pkg{manipulate} package is not installed, then an attempt is made to produce the dynamic graph with Tcl/Tk using the functions in the \pkg{relax} package.
 #' 
-#' @return None, but a dynamic graphic with slider bars is produced.
+#' @param mu0 A single numeric that is the null hypothesized mean
+#' @param s.mua A single numeric that is the starting value for the actual population mean
+#' @param s.sigma A single numeric that is the starting value for the actual population standard deviation
+#' @param s.n A single numeric that is the starting value for the sample size
+#' @param s.alpha A single numeric that is the starting value for alpha
+#' @param lower.tail A single logical that indicates if the rejection region is into the lower tail or not
+#' 
+#' @return None, but a dynamic graphic with sliders is produced.
 #' 
 #' @keywords dynamic
 #' 
@@ -284,6 +292,8 @@ iCLTSimPlot <- function(n,sh1,sh2,reps,incl.norm) {
 #' @export
 #' 
 powerSim <- function(mu0=100,s.mua=95,s.sigma=10,s.n=30,s.alpha=0.05,lower.tail=TRUE) {
+  ## Trying to fix "no visible bindings" problem for Check
+  mua <- sigma <- n <- alpha <- NULL
   if (iCheckRStudio() & requireNamespace("manipulate",quietly=TRUE)) {
     rerand <- TRUE
     manipulate::manipulate(
@@ -367,3 +377,146 @@ iPowerSimPlot <- function(mua,sigma,n,alpha,mu0,s.mua,s.sigma,s.n,lower.tail){
                    legend=paste("beta=",format(1-pwr,nsmall=3)),bty="n")
   graphics::par(old.par)
 } # end iPowerSimPlot internal function
+
+
+
+
+#' @title Dynamic simulations to demonstrate how the mean and median are computed.
+#' 
+#' @description This function visually illustrates how the mean \dQuote{balances} distances and the median \dQuote{balances} individuals in a random sample from a known population.  The user may plot their own data or compute a random sample from a beta distribution where the number of individuals in the sample and the shape of the population may be dynamically manipulated to determine how these attributes affect the calculation of the mean and median.
+#' 
+#' @details If the user is using RStudio and the \pkg{manipulate} package is installed then the dynamic graph is produced in the \dQuote{Plots} pane of RStudio.  The plot controls may be accessed through the \dQuote{gear} that is in the upper-left corner of the plot.  If the user is not using RStudio or the \pkg{manipulate} package is not installed, then an attempt is made to produce the dynamic graph with Tcl/Tk using the functions in the \pkg{relax} package.
+#' 
+#' @param x An optional numeric vector (of actual data) to be used in the visual.
+#' @param outlier A string that indicates whether the outlier should be modeled at the maximum (\code{="max"}), minimum (\code{="min"}), or not at all \code{="none"}).  This is ignored if \code{x} is not \code{NULL}.
+#' 
+#' @return None, but a graphic (if \code{x} is not \code{null}) or a dynamic graphic with sliders (if \code{x} is \code{null}) is produced.
+#' 
+#' @keywords dynamic
+#' 
+#' @examples
+#' \dontrun{
+#' ## Examples with simulated data
+#' ##   no outlier
+#' meanMedian()
+#' 
+#' ##   outlier at minimum
+#' meanMedian(outlier="min")
+#' 
+#' ##   no outlier
+#' meanMedian(outlier="max")
+#' }
+#' 
+#' ## Example with user-derived data
+#' meanMedian(c(1:7,25))
+#' 
+#' @export
+meanMedian <- function(x=NULL,outlier=c("none","max","min")) {
+  ## Trying to fix "no visible bindings" problem for Check
+  n <- shape1 <- shape2 <- NULL
+  if (!is.null(x)) {
+    ## data was supplied
+    if (!is.vector(x) | !is.numeric(x)) stop("'x' must be a numeric vector.",call.=FALSE)
+    iMMMakePlots(x)
+  } else {
+    ## Need to make up data
+    outlier <- match.arg(outlier)
+    if (iCheckRStudio() & requireNamespace("manipulate",quietly=TRUE)) {
+      rerand <- TRUE
+      manipulate::manipulate(
+        {
+          if (rerand) set.seed(sample(1:10000))
+          x <- iMMMakeData(n,shape1,shape2,outlier)
+          iMMMakePlots(x)
+        },
+        n=manipulate::slider(10,100,step=1,initial=30),
+        shape1=manipulate::slider(1,10,step=1,label="Shape 1 (alpha)"),
+        shape2=manipulate::slider(1,10,step=1,label="Shape 2 (beta)"),
+        outlier=manipulate::picker("none","min","max"),
+        rerand=manipulate::button("Rerandomize")
+      ) # end manipulate
+    } else {
+      ## internal referesher function
+      iMMBRefresh <- function(...) {
+        # get random data
+        x <- iMMMakeData(relax::slider(no=1),relax::slider(no=2),relax::slider(no=3))
+        iMMMakePlots(x)
+      } # end iMMBRefresh
+      if (iChk4Namespace("relax")) {
+        relax::gslider(iMMBRefresh,prompt=TRUE,vscale=1.5,hscale=1.75,
+                       sl.names=   c("n", "Shape 1 (alpha)", "Shape 2 (beta)"),
+                       sl.mins=    c( 10,                 1,                1),
+                       sl.maxs=    c(100,                10,               10),
+                       sl.deltas=  c(  1,                 1,                1),
+                       sl.defaults=c( 30,                 1,                1),
+                       title = "Mean vs. Median Simulator",
+                       but.functions= function(...){
+                         relax::slider(obj.name="rerand",obj.value="Y")
+                         iMMBRefresh()
+                       },
+                       but.names=c("Re-Randomize"),
+                       pos.of.panel="left")
+      }
+    }
+  }
+}
+
+
+# Internal function to make random data from a beta distribution
+iMMMakeData <- function(n,shape1,shape2,outlier) {
+  x <- stats::rbeta(n,shape1,shape2)
+  x <- x-min(x)           # rescale to a min of 0
+  x <- x/max(x)           # and a max of 1
+  if (outlier=="max") {
+    x <- x/2              # rescale max to 0.5
+    x[length(x)] <- 1     # but replace max with 1 to be an outlier
+  } else if (outlier=="min") {
+    x <- (x+1)/2          # rescale to min of 0.5 and max of 1
+    x[1] <- 0             # but replace min with 0 to be an outlier
+  }
+  x
+}
+
+## Internal function to do the plotting
+iMMMakePlots <- function(x) {
+  ## Prepare the data and some useful values
+  x <- x[order(x)]
+  mn.x <- mean(x)
+  mdn.x <- stats::median(x)
+  n <- length(x)
+  mdn.pos <- (n+1)/2
+  ## Construct the graphic
+  old.par <- graphics::par(mfcol=c(2,1))
+  # Make the histogram on top
+  graphics::par(mar=c(0.05,3.5,2,1),mgp=c(2,0.5,0),tcl=-0.2,xaxt="n")
+  graphics::hist(x,main="",col="gray90",right=FALSE)
+  graphics::abline(h=0)
+  graphics::abline(v=c(mn.x,mdn.x),col=c("red","blue"),lwd=2)
+  mdn.ttl <- paste("Median =",formatC(mdn.x,format="f",digits=2),ifelse(mdn.x>mn.x," *",""))
+  mn.ttl <- paste("Mean =",formatC(mn.x,format="f",digits=2),ifelse(mn.x>mdn.x," *",""))
+  graphics::mtext(c(mdn.ttl,mn.ttl),line=c(1,0),col=c("blue","red"))
+  # Scatterlot on the bottom
+  graphics::par(mar=c(3.5,3.5,0.05,1),mgp=c(2,0.5,0),tcl=-0.2,xaxt="s")
+  graphics::plot(x,1:n,xlab="Quantitative Variable",ylab="Ordered Individual",
+                 col.main="blue",xlim=range(pretty(range(x),n=grDevices::nclass.Sturges(x)))) 
+  # Put median on the plot
+  graphics::abline(v=mdn.x,col="blue",lwd=2)
+  if (mdn.pos == floor(mdn.pos)) { 
+    lwr <- 1:(mdn.pos-1)
+    upr <- (mdn.pos+1):n
+  } else { 
+    lwr <- 1:floor(mdn.pos)
+    upr <- (floor(mdn.pos)+1):n
+  }
+  # color points relative to the median
+  graphics::points(x[lwr],lwr,col="black",pch=19)
+  graphics::points(x[upr],upr,col="red",pch=19)  
+  # Put mean on the plot
+  graphics::abline(v=mn.x,col="red",lwd=2)
+  # put residuals from mean on graph
+  for (i in 1:n) {
+    if (x[i] < mn.x) graphics::lines(c(x[i],mn.x),c(i,i),col="black",lty=3)
+    else graphics::lines(c(x[i],mn.x),c(i,i),col="red",lty=3)
+  }
+  graphics::par(old.par)
+} # end internal iMMMakePlots